refactor: use-is-first-render로 파일 이름 변경 (#447)
