+++ src/layouts/ListLayout.tsx
@@ -11,7 +11,7 @@ import { scrollPosStore } from '@/lib/scroll-pos-store';
 import { sameAllElements } from '@/lib/utils';
 import useDebounce from '@/hooks/use-debounce';
 import useInfiniteScrollObserver from '@/hooks/use-infinite-scroll-observer';
-import useIsFirstRender from '@/hooks/useIsFirstRender';
+import useIsFirstRender from '@/hooks/use-is-first-render';
 
 import PostList from '@/components/card-and-list/post-list';
 import ScrollTopAndBottom from '@/components/common/scroll-top-and-bottom';
